http: fail when auth failure during POST
When we encounter a 401 during a POST, we must fail.  We've already sent
our content body to the remote in the initial POST request.  We did not
buffer this content body (it may be excessively large, and we may have
used chunking).  Therefore we cannot replay these requests.  Simply
fail.