gssapi: send a probe packet before a POST
SPNEGO was designed around the concept of authenticating an entire
connection, or socket.  By taking advantage of HTTP keep-alive, an
entire socket is authenticated once.  Subsequent calls over the same
connection will maintain the same authentication context.  This reduces
the number of challenge/response calls that need to be made.

Apache, however, doesn't do this.

It requires re-authentication for every individual HTTP request, as if
it were Basic or Digest.  This gives terrible performance for GET
requests, but worse: for POST requests, it requires users to re-send the
POST data.

So when we're dealing with NTLM or SPNEGO authentication, send a POST
with a no-op "probe packet".  If we have authenticated previously, and
we're talking to a server that understands connection affinity (ie, IIS)
then this will give us back a 200 and simply waste a round-trip.  But
Apache servers will give us a 401 with the authentication challenge.
We'll take that, compute our response, and then send a new POST packet
with our _actual_ data.

This allows us to authenticate during the POST without sending our
actual POST data, which is not buffered or rewindable.